Remove the import from core to components/ActivityCollector
